generate.ts to index.ts
